requirements.txt
crispy-bootstrap4==2022.1
Django==4.0.3
django-chart-js==0.0.3
django-chartjs==2.3.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.5
Jinja2==3.1.2
matplotlib==3.7.0
numpy==1.24.1
openpyxl==3.0.10
pytest==7.2.1
pyunpack==0.2.2
requests==2.28.1
schedule==1.2.0
urllib3==1.26.14
xlrd==2.0.1
xls2xlsx==0.2.0
django-debug-toolbar==4.3.0
lasio
python-pptx
tqdm